Algorithm_material_system/requirements.txt
# 基础依赖
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scikit-learn>=0.23.0
torch>=1.7.0
tqdm>=4.50.0
seaborn>=0.11.0
Pillow>=8.0.0

# API依赖
Flask>=2.0.0
Flask-Cors>=3.0.10
waitress>=2.0.0
gunicorn>=20.1.0  # 用于Linux系统部署
requests>=2.25.0

# 工具依赖
chardet>=4.0.0
